Add IndexerId trait, indexer_network_subgraph_entry
